+++ vsd/src/hls/playlist.rs
@@ -1,4 +1,5 @@
 use crate::playlist;
+use url::Url;
 
 pub(crate) fn parse_as_master(
     m3u8: &m3u8_rs::MasterPlaylist,
@@ -25,7 +26,7 @@ pub(crate) fn parse_as_master(
                 None
             },
             segments: vec![], // Cannot be comment here
-            uri: video_stream.uri.to_owned(),
+            uri: video_stream.uri.parse::<Url>().unwrap(),
         });
     }
 
@@ -45,7 +46,7 @@ pub(crate) fn parse_as_master(
                     playlist_type: playlist::PlaylistType::Hls,
                     resolution: None, // Cannot be comment here
                     segments: vec![], // Cannot be comment here
-                    uri: uri.to_owned(),
+                    uri: uri.parse::<Url>().unwrap(),
                 }),
 
                 m3u8_rs::AlternativeMediaType::Audio => streams.push(playlist::MediaPlaylist {
@@ -67,7 +68,7 @@ pub(crate) fn parse_as_master(
                     playlist_type: playlist::PlaylistType::Hls,
                     resolution: None,
                     segments: vec![], // Cannot be comment here
-                    uri: uri.to_owned(),
+                    uri: uri.parse::<Url>().unwrap(),
                 }),
 
                 m3u8_rs::AlternativeMediaType::ClosedCaptions
@@ -88,7 +89,7 @@ pub(crate) fn parse_as_master(
                         playlist_type: playlist::PlaylistType::Hls,
                         resolution: None,
                         segments: vec![], // Cannot be comment here
-                        uri: uri.to_owned(),
+                        uri: uri.parse::<Url>().unwrap(),
                     })
                 }
 
@@ -111,7 +112,7 @@ pub(crate) fn parse_as_master(
                     playlist_type: playlist::PlaylistType::Hls,
                     resolution: None, // Cannot be comment here
                     segments: vec![], // Cannot be comment here
-                    uri: uri.to_owned(),
+                    uri: uri.parse::<Url>().unwrap(),
                 }),
             }
         }
@@ -132,7 +133,7 @@ pub(crate) fn push_segments(m3u8: &m3u8_rs::MediaPlaylist, playlist: &mut playli
 
     for segment in &m3u8.segments {
         let map = segment.map.as_ref().map(|x| playlist::Map {
-            uri: x.uri.to_owned(),
+            uri: x.uri.parse::<Url>().unwrap(),
             range: x.byte_range.as_ref().map(|x| {
                 let offset = x.offset.unwrap_or(0);
 
@@ -210,19 +211,19 @@ pub(crate) fn push_segments(m3u8: &m3u8_rs::MediaPlaylist, playlist: &mut playli
             },
             map,
             range,
-            uri: segment.uri.to_owned(),
+            uri: segment.uri.parse::<Url>().unwrap(),
         });
     }
 
     if let Some(segment) = playlist.segments.get(0) {
         if let Some(init) = &segment.map {
-            if init.uri.split('?').next().unwrap().ends_with(".mp4") {
+            if init.uri.as_str().split('?').next().unwrap().ends_with(".mp4") {
                 playlist.extension = Some("m4s".to_owned());
             }
         }
 
         if let Some(extension) = segment
-            .uri
+            .uri.as_str()
             .split('?')
             .next()
             .unwrap()